Fix - Utilisation d'un decimal pour le FixedAmount

+++ Sheaft.Options/PspOptions.cs
@@ -11,7 +11,7 @@ namespace Sheaft.Options
         public string PaylineUrl { get; set; }
         public string ClientId { get; set; }
         public string ApiKey { get; set; }
-        public int FixedAmount { get; set; }
+        public decimal FixedAmount { get; set; }
         public decimal Percent { get; set; }
         public string Token
         {
